Aqua/refactor - reorganize code (#837)
