+++ ads/aqua/extension/deployment_handler.py
@@ -7,9 +7,10 @@ from urllib.parse import urlparse
 
 from tornado.web import HTTPError
 
-from ads.aqua.decorator import handle_exceptions
-from ads.aqua.deployment import AquaDeploymentApp, MDInferenceResponse, ModelParams
+from ads.aqua.common.decorator import handle_exceptions
 from ads.aqua.extension.base_handler import AquaAPIhandler, Errors
+from ads.aqua.modeldeployment import AquaDeploymentApp, MDInferenceResponse
+from ads.aqua.modeldeployment.entities import ModelParams
 from ads.config import COMPARTMENT_OCID, PROJECT_OCID
