+++ ads/aqua/extension/finetune_handler.py
@@ -8,10 +8,11 @@ from urllib.parse import urlparse
 
 from tornado.web import HTTPError
 
-from ads.aqua.decorator import handle_exceptions
+from ads.aqua.common.decorator import handle_exceptions
 from ads.aqua.extension.base_handler import AquaAPIhandler, Errors
 from ads.aqua.extension.utils import validate_function_parameters
-from ads.aqua.finetune import AquaFineTuningApp, CreateFineTuningDetails
+from ads.aqua.finetuning import AquaFineTuningApp
+from ads.aqua.finetuning.entities import CreateFineTuningDetails
 
 
 class AquaFineTuneHandler(AquaAPIhandler):
