+++ ads/aqua/finetuning/__init__.py
@@ -1,0 +1,7 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+# Copyright (c) 2024 Oracle and/or its affiliates.
+# Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl/
+from ads.aqua.finetuning.finetuning import AquaFineTuningApp
+
+__all__ = ["AquaFineTuningApp"]
