+++ ads/aqua/modeldeployment/__init__.py
@@ -1,0 +1,8 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+# Copyright (c) 2024 Oracle and/or its affiliates.
+# Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl/
+from ads.aqua.modeldeployment.deployment import AquaDeploymentApp
+from ads.aqua.modeldeployment.inference import MDInferenceResponse
+
+__all__ = ["AquaDeploymentApp", "MDInferenceResponse"]
