+++ tests/unitary/with_extras/aqua/test_finetuning_handler.py
@@ -6,14 +6,16 @@
 
 from unittest import TestCase
 from unittest.mock import MagicMock
-from mock import patch
 
-from notebook.base.handlers import IPythonHandler, APIHandler
+from mock import patch
+from notebook.base.handlers import APIHandler, IPythonHandler
+
 from ads.aqua.extension.finetune_handler import (
     AquaFineTuneHandler,
     AquaFineTuneParamsHandler,
 )
-from ads.aqua.finetune import AquaFineTuningApp, CreateFineTuningDetails
+from ads.aqua.finetuning import AquaFineTuningApp
+from ads.aqua.finetuning.entities import CreateFineTuningDetails
 
 
 class TestDataset:
